Catch up with January 2016 develop
Caught bortsc version up with DRI's January 2016 changes to develop
version.
