--- docs/run_basin.rst
@@ -1,2 +1,0 @@
-Run Basin
-=========
