+++ et-demands/prep/build_ag_cdl_rasters.py
@@ -2,7 +2,7 @@
 # Name:         build_ag_cdl_rasters.py
 # Purpose:      Build agricultural land and mask rasters from CDL rasters
 # Author:       Charles Morton
-# Created       2016-07-22
+# Created       2017-01-11
 # Python:       2.7
 #--------------------------------
 
@@ -16,8 +16,8 @@ import sys
 import numpy as np
 from osgeo import gdal
 
-import gdal_common as gdc
-import util
+import _gdal_common as gdc
+import _util as util
 
 
 def main(gis_ws, cdl_year='', block_size=16384, mask_flag=False,
@@ -51,9 +51,9 @@ def main(gis_ws, cdl_year='', block_size=16384, mask_flag=False,
 
     # Ag landuses are 1, all others in state are 0, outside state is nodata
     # Crop 61 is fallow/idle and was excluded from analysis
-    # Crops 176 is Grassland/Pasture in the new national CDL rasters
-    # Crops 181 was Pasture/Hay in the old state CDL rasters
-    # Crops 182 was Cultivated Crop in the old state CDL rasters
+    # Crop 176 is Grassland/Pasture in the new national CDL rasters
+    # Crop 181 was Pasture/Hay in the old state CDL rasters
+    # Crop 182 was Cultivated Crop in the old state CDL rasters
     agmask_remap = [
         [1, 60, 1], [61, 65, 0],
         # [1, 61, 1], [62, 65, 0],
