+++ et-demands/prep/build_spatial_crop_params_arcpy.py
@@ -2,7 +2,7 @@
 # Name:         build_spatial_crop_params_arcpy.py
 # Purpose:      Build spatial parameter files for ET-Demands from zonal stats ETCells
 # Author:       Charles Morton
-# Created       2016-07-22
+# Created       2017-01-11
 # Python:       2.7
 #--------------------------------
 
@@ -16,7 +16,7 @@ import sys
 
 import arcpy
 
-import util
+import _util as util
 
 
 def main(ini_path, zone_type='huc8', area_threshold=10,
