+++ example/cet_example.ini
@@ -10,10 +10,9 @@ cells_path = D:\et-demands\example\gis\ETCells.shp
 crop_et_folder = D:\et-demands\et-demands\cropET
 template_folder = D:\et-demands\et-demands\static
 
+## Crops
 # crop_test_list = 3, 7
 # crop_skip_list = 55-57
-# annual_skip_flag = True
-# perennial_skip_flag = False
 
 ## Stats flags
 daily_stats_flag = True
@@ -23,7 +22,6 @@ growing_season_stats_flag = True
 
 ## Spatially varying calibration
 spatial_cal_flag = False
-spatial_cal_folder = D:\et-demands\example\calibration_shapefiles
 
 ## Output alfalfa cuttings
 cutting_flag = True
@@ -31,17 +29,15 @@ cutting_flag = True
 niwr_flag = True
 ## Compute crop coefficient (Kc)
 kc_flag = True
+
 ## CO2 correction
 co2_flag = False
-# co2_grass_list = 1-6,9-18, 21-67, 69, 71-73, 75, 79-81, 83-85
-# co2_trees_list = 19, 20, 70, 74, 82
-# co2_c4_list = 7, 8, 68, 76-78
 
 ## Limit to a date range (YYYY-MM-DD)
 start_date = None
 end_date = None
 
-## ET sub-folder names
+## Sub folder names
 static_folder = static
 daily_output_folder = daily_stats
 monthly_output_folder = monthly_stats
