+++ example/csv_cet_dri.ini
@@ -207,7 +207,7 @@ kc_flag = True
 
 co2_flag = False
 co2_grass_list = 1-6,9-18, 21-67, 69, 71-73, 75, 79-81, 83-85
-co2_trees_list = 19, 20, 70, 74, 82
+co2_tree_list = 19, 20, 70, 74, 82
 co2_c4_list = 7, 8, 68, 76-78
 
 # cet output folder names
