+++ requirements.txt
@@ -2,4 +2,4 @@ bokeh>=0.12.0
 gdal>=2.0
 numpy
 pandas
-openpyxl>=2.4.0-a1
+openpyxl>=2.4.7
